Extract tutorial from NewsletterView and SearchResults into reusable component
