+++ src/components/NewsletterView.js
@@ -1,61 +1,13 @@
 import React from 'react';
 
 import Card from './Card';
+import Tutorial from './Tutorial';
 import { withNewsletter } from '../util/NewsletterContext';
 
 import './NewsletterView.css';
 
 const downloadURL = item =>
   `https://phaser.io/images/newsletter/pdf/issue${item.Issue}.pdf`;
-
-const Tutorial = ({ tutorial }) => {
-  return (
-    <li className="tutorial" aria-label="Tutorial list item">
-      <Card
-        header={
-          <>
-            <h4 className="card-title">{tutorial.name}</h4>
-            <div
-              className="card-badge"
-              aria-label="Phaser version for this tutorial"
-            >
-              Phaser {tutorial.version === 'v2' ? '2/CE' : '3'}
-            </div>
-          </>
-        }
-        content={<p aria-label="Tutorial description">{tutorial.desc}</p>}
-        footer={
-          <div className="tags" aria-label="Tags">
-            {tutorial.tags &&
-              tutorial.tags.map(tag => {
-                return (
-                  <div className="card-badge" key={tag}>
-                    {tag}
-                  </div>
-                );
-              })}
-          </div>
-        }
-        links={
-          <>
-            <a href={tutorial.link} target="_blank" rel="noopener noreferrer">
-              Read more on phaser.io
-            </a>
-            {tutorial.directlink && (
-              <a
-                href={tutorial.directlink}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                Go directly to site
-              </a>
-            )}
-          </>
-        }
-      />
-    </li>
-  );
-};
 
 class NewsletterView extends React.Component {
   constructor(props) {
